Impementation of COUNTRY-LEVEL DamFrac BUT no aggregation yet
Country-level DAMFRAC (% terms) is working.
Country-level DAMAGES (absolute) not implemented yet (and I'm not sure whether this is needed)
Thus, country-level damages are not aggregated yet (also not sure whether this is needed, might just give me the same results as the region level damages). So the model still runs with regional DAMAGES.