Dropping the "^10" in the damage function
i.e. in the translation of damage fractions to damages in absolute terms

Reason: James' recommendation, because:

"The point of the expression with damages / (1 + damages^10) is that it will ensure that damages don't exceed 100%, though you would want to apply it on at the country level. But I don't like that expression, because it results in marginally decreasing corrected damages for high total damages. Better to just drop that ^10."